TA help with Code
Per guidance from the TA, this python code runs an "/is-odd" route on the flask program.  He provides other info as follows: 

So you see the "@app.route('/')" proclamation, it's basically the routing for the flask application. Meaning that when a user goes to http:localhost:8080\ it defaults to saying Hello with the H1 tag
So what's next is to define the other routes for our page, expose the URIs
So for example if I wanted an /is-odd route"